♻️ Refactor: Move queue card to separate folder
